Refactored 'gather' module to make import path a little cleaner
