+++ tests/unit_tests/gather_tests/MongoSink_test.py
@@ -2,7 +2,7 @@ import json
 import pytest
 import msgpack
 import pymongo
-from cps2_zmq.gather.MongoSink import MongoSink
+from cps2_zmq.gather import MongoSink
 from cps2_zmq.gather import mdp
 
 idn = 'sink-1'
@@ -10,7 +10,6 @@ port = 6666
 addr = ':'.join(["tcp://127.0.0.1", str(port)])
 sub_addr = ':'.join(["tcp://127.0.0.1", str(port + 1)])
 db_name = 'sink_test'
-
 
 @pytest.fixture(scope='function')
 def sink():
@@ -26,6 +25,7 @@ def db():
     client.drop_database(db_name)
     client.close()
 
+@pytest.mark.skip
 def test_process_pub(sink, db):
     message = json.dumps({'test': 'message'})
     sink.process_pub(msgpack.packb(message))
